cors config prod

+++ ezworks-backend/src/main/java/com/ezworks/repository/RolRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RolRepository extends JpaRepository<Rol, Short> {
+public interface RolRepository extends JpaRepository<Rol, Byte> {
 
     Optional<Rol> findByCodigo(RolCodigo codigo);
 }
